+++ 13-event-emitter.js
@@ -1,0 +1,17 @@
+const EventEmitter = require("events");
+
+const customEmitter = new EventEmitter();
+
+//on and emit methods 
+//keep track of the order
+//additional arguments
+//built-in modules utilize it
+
+customEmitter.on("response", (name, id) => {
+  console.log(`data recieved user ${name} with id: ${id}`);
+});
+customEmitter.on("response", () => {
+  console.log(`some other logic here `);
+});
+
+customEmitter.emit("response", 'ximi', 0345);
