+++ app.js
@@ -1,40 +1,15 @@
-const { readFile, writeFile } = require("fs").promises;
+const http = require("http");
 
-// const util = require("util");
-// const readFilePromise = util.promisify(readFile);
-// const writeFilePromise = util.promisify(writeFile);
+// const server = http.createServer((req,res) => {
+//     res.end('Welcome')
+// })
 
-// const getText = (path) => {
-//   return new Promise((resolve, reject) => {
-//     readFile(path, "utf8", (err, data) => {
-//       if (err) {
-//         reject(err);
-//       } else {
-//         resolve(data);
-//       }
-//     });
-//   });
-// };
+//Using Event Emitter API
+const server = http.createServer();
+//emits request event
+//subscribe to it/ liten for it/ respond to it
+server.on("request", (req, res) => {
+  res.end("Welcome");
+});
 
-// getText("./content/first.txt")
-//   .then((result) => console.log(result))
-//   .catch((err) => {
-//     console.log(err);
-//   });
-
-const start = async () => {
-  try {
-    const first = await readFile("./content/first.txt", "utf-8");
-    const second = await readFile("./content/second.txt", "utf-8");
-    await writeFile(
-      "./content/result-mind-granade.txt",
-      ` THIS IS AWESOME ${first} ${second}`,
-      { flag: 'a' }
-    );
-    console.log(first, second);
-  } catch (error) {
-    console.log(error);
-  }
-};
-
-start();
+server.listen(5000);
